Sink implementation of SourceFile::getIfConfigClauseRanges into Sema
This implementation depends on ASTGen, which isn't linked as part of the
AST library.
